debug and space modules added
