+++ lib/src/metta/runner/stdlib/debug.rs
@@ -1,0 +1,112 @@
+use crate::*;
+use crate::metta::*;
+use crate::metta::text::Tokenizer;
+
+use crate::metta::runner::stdlib::{grounded_op, atom_to_string, regex};
+
+use std::convert::TryInto;
+
+
+/// Implement trace! built-in.
+///
+/// It is equivalent to Idris or Haskell Trace, that is, it prints a
+/// message to stderr and pass a value along.
+///
+/// For instance
+/// ```metta
+/// !(trace! "Here?" 42)
+/// ```
+/// prints to stderr
+/// ```stderr
+/// Here?
+/// ```
+/// and returns
+/// ```metta
+/// [42]
+/// ```
+///
+/// Note that the first argument does not need to be a string, which
+/// makes `trace!` actually quite capable on its own.  For instance
+/// ```metta
+/// !(trace! ("Hello world!" (if True A B) 1 2 3) 42)
+/// ```
+/// prints to stderr
+/// ```stderr
+/// (Hello world! A 1 2 3)
+/// ```
+/// and returns
+/// ```metta
+/// [42]
+/// ```
+
+#[derive(Clone, Debug)]
+pub struct TraceOp {}
+
+grounded_op!(TraceOp, "trace!");
+
+impl Grounded for TraceOp {
+    fn type_(&self) -> Atom {
+        Atom::expr([ARROW_SYMBOL, ATOM_TYPE_UNDEFINED, Atom::var("a"), Atom::var("a")])
+    }
+
+    fn as_execute(&self) -> Option<&dyn CustomExecute> {
+        Some(self)
+    }
+}
+
+impl CustomExecute for TraceOp {
+    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
+        let arg_error = || ExecError::from("trace! expects two atoms as arguments");
+        let val = args.get(1).ok_or_else(arg_error)?;
+        let msg = args.get(0).ok_or_else(arg_error)?;
+        eprintln!("{}", msg);
+        Ok(vec![val.clone()])
+    }
+}
+
+#[derive(Clone, Debug)]
+pub struct PrintAlternativesOp {}
+
+grounded_op!(PrintAlternativesOp, "print-alternatives!");
+
+impl Grounded for PrintAlternativesOp {
+    fn type_(&self) -> Atom {
+        Atom::expr([ARROW_SYMBOL, ATOM_TYPE_ATOM, ATOM_TYPE_EXPRESSION, UNIT_TYPE])
+    }
+
+    fn as_execute(&self) -> Option<&dyn CustomExecute> {
+        Some(self)
+    }
+}
+
+impl CustomExecute for PrintAlternativesOp {
+    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
+        let arg_error = || ExecError::from("print-alternatives! expects format string as a first argument and expression as a second argument");
+        let atom = atom_to_string(args.get(0).ok_or_else(arg_error)?);
+        let args = TryInto::<&ExpressionAtom>::try_into(args.get(1).ok_or_else(arg_error)?)?;
+        let args: Vec<String> = args.children().iter()
+            .map(|atom| atom_to_string(atom))
+            .collect();
+        println!("{} {}:", args.len(), atom);
+        args.iter().for_each(|arg| println!("    {}", arg));
+        Ok(vec![UNIT_ATOM])
+    }
+}
+
+pub fn register_runner_tokens(tref: &mut Tokenizer) {
+    let trace_op = Atom::gnd(TraceOp{});
+    tref.register_token(regex(r"trace!"), move |_| { trace_op.clone() });
+    let print_alternatives_op = Atom::gnd(PrintAlternativesOp{});
+    tref.register_token(regex(r"print-alternatives!"), move |_| { print_alternatives_op.clone() });
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn trace_op() {
+        assert_eq!(TraceOp{}.execute(&mut vec![sym!("\"Here?\""), sym!("42")]),
+                   Ok(vec![sym!("42")]));
+    }
+}
